[Initial POC - Read DA Info]
Started a fresh solution to migrate my code over and separate out into a console application and a GUI. Currently has the library to parse the difficulty adjustment file for damage values and a windows form to simply load said file.

The file it targets is the "gamerankparameterdata.user.2" file found in the pak data of RE3R.

Right now, just can parse the value and spit them out into the debugger of Visual Studio.