[Correction of DA Rank interpretation + UI Simplification]
Realised after testing using the Speedrun Tool available for RE3R I was misinterpreting the Data Values for the DA system.

Corrected that now, and it led to a much more simple UI for the user to modify the damage values now.

Also added a little side panel in case new modders stumble onto this and want an explanation of what it is for.